requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
requests==2.32.3
openai==1.59.7
pydantic==2.10.4
moviepy==1.0.3
pillow==10.4.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
